FMD6
  1. Fixing Bugzila BUGs 4143, 3482, 4090, 4074, 4230, 3182, 4052, 3935, 4413, 4450
  2. Added workarounds for FM chip erratas.
  3. Fixed several BUGs found during bring-up
  4. Added support for sw-parser

+++ drivers/net/dpa/NetCommSw/user/env/linux/kernel/2.6/inc/wrappers/Peripherals/fsl_fman.h
@@ -106,7 +106,7 @@ struct fm_port_rx_params {
     uint8_t                     num_pools;          /**< Number of pools use by this port */
     struct fm_port_pool_param   pool_param[FM_PORT_MAX_NUM_OF_EXT_POOLS];
                                                     /**< Parameters for each pool */
-    uint8_t                     priv_data_size;     /**< Area that user may save for his own need (E.g. save the SKB) */
+    uint16_t                    priv_data_size;     /**< Area that user may save for his own need (E.g. save the SKB) */
     bool                        parse_results;      /**< Whether to have the parser-results in the Received buffer */
 };
 
@@ -118,7 +118,7 @@ struct fm_port_non_rx_params {
     uint32_t                defq;               /**< For Tx and HC - Default Confirmation queue,
                                                      0 means no Tx confirmation for processed
                                                      frames. For OP - default Rx queue. */
-    uint8_t                 priv_data_size;     /**< Area that user may save for his own need (E.g. save the SKB) */
+    uint16_t                priv_data_size;     /**< Area that user may save for his own need (E.g. save the SKB) */
     bool                    parse_results;      /**< Whether to put the parser-results in the Transmitted buffer */
 };
